factor in_between_test functions into files
