+++ src/interpolation/interp_manner.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+from typing import Callable, TypeAlias
+from .interp_strategy import InterpStrategy
+
+QuotientModifier: TypeAlias = Callable[[float], float]
+
+@dataclass
+class InterpManner:
+  quotient_modifier: QuotientModifier
+  strategy: InterpStrategy
