+++ src/sample_spec/sample_spec.py
@@ -1,0 +1,5 @@
+from typing import TypeAlias, Union
+from .prompts import PromptType
+from ..interpolation.inter_prompt import InterPrompt
+
+SampleSpec: TypeAlias = Union[PromptType, InterPrompt[PromptType]]
